Allow libFLAC to be shared (or static)
By changing the cc_library_static to cc_library, it can be included as a
static_library or shared_library as needed.

In order to allow the IAMF decoder and the FLAC decoder to share the
library, this CL makes it possible for the library to be included as
shared.

Bug: 396187214
Flag: android.media.swcodec.flags.iamf_aac_flac
Test: build

Change-Id: I7fbeee568c90cc1ed4430a28a5450533cb62afd3